requirements.txt
click>=8.0.0
pytest>=7.0.0
graphviz>=0.20.0
pyyaml>=6.0.0
